Fix errors related to number parsing

+++ app/src/main/java/com/example/crowdfundingplatform/data/remote/model/UserModel.kt
@@ -1,5 +1,7 @@
 package com.example.crowdfundingplatform.data.remote.model
 
+import java.math.BigDecimal
+
 data class UserModel(
     val id: String,
     val personRole: PersonRole?,
@@ -7,7 +9,7 @@ data class UserModel(
     val surname: String?,
     val patronymic: String?,
     val email: String?,
-    val money: Int?,
+    val money: BigDecimal?,
     val bio: String?,
     val emailIsConfirm: Boolean?,
     val avatarId: String?
