+++ app/src/main/java/com/example/crowdfundingplatform/domain/entity/user/User.kt
@@ -1,5 +1,7 @@
 package com.example.crowdfundingplatform.domain.entity.user
 
+import java.math.BigDecimal
+
 data class User(
     val id: String,
     val userRole: UserRole,
@@ -7,7 +9,7 @@ data class User(
     val surname: String,
     val patronymic: String,
     val email: String,
-    val money: Int,
+    val money: BigDecimal,
     val bio: String,
     val emailIsConfirmed: Boolean,
     val avatarId: String?
